dedupe-tile-component * Create components out of BusStopTile (Tile) and BusStopHeader (TileHeader)
