+++ src/components/Tile/Tile.module.scss
@@ -1,0 +1,10 @@
+.tile {
+    width: 100%;
+    max-height: 80vh;
+    padding: 2rem;
+    margin-bottom: 2rem;
+    overflow: scroll;
+    border-radius: 1rem;
+    background-color: var(--tavla-box-background-color);
+    color: var(--tavla-font-color);
+}
