+++ src/dashboards/BusStop/components/BusStopTile/BusStopTile.tsx
@@ -7,7 +7,6 @@ import {
 } from '../../../../types'
 import { useSettings } from '../../../../settings/SettingsProvider'
 import { WalkInfo } from '../../../../logic/use-walk-info/useWalkInfo'
-import { BusStopHeader } from '../BusStopHeader/BusStopHeader'
 import { BusStopTableRow } from '../BusStopTableRow/BusStopTableRow'
 import { isNotNullOrUndefined } from '../../../../utils/typeguards'
 import { unique } from '../../../../utils/array'
@@ -16,6 +15,8 @@ import {
     getIconColorType,
     getTransportIconIdentifier,
 } from '../../../../utils/icon'
+import { Tile } from '../../../../components/Tile/Tile'
+import { TileHeader } from '../../../../components/TileHeader/TileHeader'
 import css from './BusStopTile.module.scss'
 
 function getTransportHeaderIcons(
@@ -49,8 +50,8 @@ const BusStopTile = ({
     )
 
     return (
-        <div className={css.busStopTile}>
-            <BusStopHeader
+        <Tile>
+            <TileHeader
                 title={name}
                 icons={getTransportHeaderIcons(departures, iconColorType)}
                 walkInfo={!settings.hideWalkInfo ? walkInfo : undefined}
@@ -87,7 +88,7 @@ const BusStopTile = ({
                     ))}
                 </TableBody>
             </Table>
-        </div>
+        </Tile>
     )
 }
